Remove iframe injection for GTM
There's probably no point in injecting the GTM iframe here. As per the [documentation](https://developers.google.com/tag-platform/tag-manager/web), this iframe is meant to be enclosed within a `<noscript>` tag so that GTM loads for visitors who have disabled Javascript. As this script runs in the browser anyways, it implies that it executes only for visitors who have Javascript enabled, so the iframe is redundant, and actually results in an extra network call.